+++ include/ygm/random.hpp
@@ -5,55 +5,14 @@
 
 #pragma once
 
-#include <ygm/comm.hpp>
-
-#include <random>
+#include <ygm/detail/random.hpp>
 
 namespace ygm {
-namespace random {
-
-/// @brief Applies a simple offset to the specified seed according to rank index
-/// @tparam ResultType The random number (seed) type
-/// @param comm The ygm::comm to be used
-/// @param seed The specified seed
-/// @return simply returns seed + rank
-template <typename ResultType>
-ResultType simple_offset(ygm::comm &comm, ResultType seed) {
-  return seed + comm.rank();
-}
-
-/// @brief A wrapper around a per-rank random engine that manipulates each
-///        rank's seed according to a specified strategy
-/// @tparam RandomEngine The underlying random engine, e.g. std::mt19337
-/// @tparam Function A function on (ygm::comm, result_type) -> result_type that
-///         modifies seeds for each rank
-template <typename RandomEngine,
-          typename RandomEngine::result_type (*Function)(
-              ygm::comm &, typename RandomEngine::result_type)>
-class random_engine {
- public:
-  using rng_type    = RandomEngine;
-  using result_type = typename RandomEngine::result_type;
-
-  random_engine(ygm::comm &comm, result_type seed = std::random_device{}())
-      : m_seed(Function(comm, seed)), m_rng(Function(comm, seed)) {}
-
-  result_type operator()() { return m_rng(); }
-
-  constexpr const result_type &seed() const { return m_seed; }
-
-  static constexpr result_type min() { return rng_type::min(); }
-  static constexpr result_type max() { return rng_type::max(); }
-
- private:
-  rng_type    m_rng;
-  result_type m_seed;
-};
 
 /// @brief A simple offset rng alias
-/// @tparam RandomEngine The underlying random engine, e.g. std::mt19337
-template <typename RandomEngine>
-using default_random_engine = random_engine<RandomEngine, simple_offset>;
+/// @tparam RandomEngine The underlying random engine, e.g. std::mt19937
+template <typename RandomEngine = std::mt19937>
+using default_random_engine =
+    ygm::detail::random_engine<RandomEngine, ygm::detail::simple_offset>;
 
-}  // namespace random
 }  // namespace ygm
