+++ test/test_random.cpp
@@ -16,11 +16,11 @@ int main(int argc, char** argv) {
   //
   // Test default_random_engine
   {
-    std::uint32_t                                    seed = 100;
-    ygm::random::default_random_engine<std::mt19937> rng(world, seed);
-    ygm::container::counting_set<int>                seed_set(world);
-    ygm::container::counting_set<int>                rn_set(world);
-    ygm::container::counting_set<int>                sample_set(world);
+    std::uint32_t                     seed = 100;
+    ygm::default_random_engine<>      rng(world, seed);
+    ygm::container::counting_set<int> seed_set(world);
+    ygm::container::counting_set<int> rn_set(world);
+    ygm::container::counting_set<int> sample_set(world);
 
     std::uint32_t                                local_rn = rng();
     std::uniform_int_distribution<std::uint32_t> dist(0, 10000000);
